Add component refactoring for improved organization and maintainability. Created new files for agent diary and refactoring plan documentation. Extracted components from landing and dashboard pages into logical groupings, enhancing code clarity and separation of concerns. Updated page files to utilize new components, reducing complexity and improving maintainability.
