+++ agent-sessions/refactor-components-2025-08-08/agent-diary.md
@@ -1,0 +1,29 @@
+# Agent Diary - Component Refactoring Session
+Date: 2025-08-08
+
+## Session Overview
+Refactoring the codebase to improve component organization and reduce page file complexity.
+
+## Progress Log
+
+### 09:00 - Session Start
+- Read initialization protocols from `/docs/protocols/init.md` and `/docs/protocols/start-work.md`
+- Reviewed current component structure in `/components/` directory
+- Analyzed `/app/page.tsx` (landing page) and `/app/dashboard/page.tsx` (dashboard) for refactoring opportunities
+
+### Current State Analysis
+1. **Components Directory**: Has UI components from shadcn/ui and some custom components, but lacks logical grouping
+2. **Landing Page (`/app/page.tsx`)**: 194 lines, contains inline component definitions and sections
+3. **Dashboard Page (`/app/dashboard/page.tsx`)**: 459 lines, very large file with multiple inline components
+
+### Identified Issues
+- Large monolithic page components
+- Inline component definitions within pages
+- No logical grouping of related components
+- Missing separation of concerns
+
+## Next Steps
+1. Create logical component groups in `/components/` directory
+2. Extract inline components from pages into separate files
+3. Refactor page files to use imported components
+4. Improve code organization and maintainability
